pyproject.toml
[project]
name = "python-fastapi-postgres-boilerplate"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116",
  "uvicorn[standard] (>=0.35.0,<0.36.0)",
  "pydantic[email] (>=2.11.7,<3.0.0)",
  "SQLAlchemy>=2.0",
  "asyncpg>=0.30",
  "alembic>=1.16",
  "passlib[bcrypt] (>=1.7.4,<2.0.0)",
  "PyJWT>=2.10",
  "python-dotenv>=1.1",
  "psycopg-binary (>=3.2.9,<4.0.0)",
  "email-validator (>=2.3.0,<3.0.0)",
  "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]

